pyproject.toml
[project]
name = "fundos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "boto3>=1.38.3",
    "colorlog>=6.9.0",
    "email-validator>=2.2.0",
    "fastapi>=0.115.12",
    "greenlet>=3.2.1",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.3",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "python-json-logger>=3.3.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.40",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.34.2",
]
